Replaced all curvedata files with extended versions
These curvedata files contain extra columns for the data which is in
intpts, alldegphi, galrep and 2adic files.  After this the only files
needed for the LMFDB are curvedata and classdata files.  We will keep
the separate intpts, alldegphi, galrep and 2adic files since they are
listed on and available for download from the ecdata website.